+++ packages/ui/src/pages/Profile/Accounts.tsx
@@ -1,15 +1,14 @@
-import Identicon from '@polkadot/react-identicon'
 import React from 'react'
 import styled from 'styled-components'
+import BN from 'bn.js'
 import { ButtonGhostMediumSquare } from '../../components/buttons/Buttons'
-import { CopyButton } from '../../components/buttons/CopyButton'
 import { ArrowDownIcon } from '../../components/icons/ArrowDownIcon'
 import { ArrowInsideIcon } from '../../components/icons/ArrowInsideIcon'
 import { TransferButton } from '../../components/TransferButton'
+import { AccountInfo } from '../../components/AccountInfo'
 import { BorderRad, Colors, Transitions } from '../../constants'
 import { useAccounts } from '../../hooks/useAccounts'
 import { useBalances } from '../../hooks/useBalances'
-import BN from 'bn.js'
 import { formatTokenValue } from '../../utils/formatters'
 
 export function Accounts() {
@@ -38,17 +37,7 @@ export function Accounts() {
           <AccountsList>
             {allAccounts.map((account) => (
               <AccountItem key={account.address}>
-                <AccountInfo>
-                  <AccountPhoto>
-                    <Identicon size={40} theme={'beachball'} value={account.address} />
-                  </AccountPhoto>
-                  {/*<AccountType>Root account</AccountType>*/}
-                  <AccountName>{account.name}</AccountName>
-                  <AccountCopyAddress>
-                    <AccountAddress>{account.address}</AccountAddress>
-                    <AccountCopyButton />
-                  </AccountCopyAddress>
-                </AccountInfo>
+                <AccountInfo account={account} />
                 <AccountBalance>
                   <Balance value={balances.map[account.address]?.total} />
                 </AccountBalance>
@@ -225,79 +214,6 @@ const AccountItem = styled.li`
   padding: 16px 0 16px 14px;
   border: 1px solid ${Colors.Black[100]};
   border-radius: ${BorderRad.s};
-`
-
-const AccountInfo = styled.div`
-  display: grid;
-  grid-template-columns: 40px 1fr;
-  grid-template-rows: min-content, 24px 18px;
-  grid-column-gap: 12px;
-  grid-template-areas:
-    'accountphoto accounttype'
-    'accountphoto accountname'
-    'accountphoto accountaddress';
-  align-items: center;
-  width: 100%;
-  justify-self: start;
-`
-
-const AccountPhoto = styled.div`
-  display: flex;
-  grid-area: accountphoto;
-  justify-content: flex-end;
-  align-items: center;
-  align-content: center;
-  align-self: center;
-  height: 40px;
-  width: 40px;
-  background-color: ${Colors.Blue[500]};
-  border-radius: ${BorderRad.full};
-  overflow: hidden;
-`
-
-// const AccountType = styled.p`
-//   display: flex;
-//   grid-area: accounttype;
-//   justify-content: center;
-//   width: fit-content;
-//   margin: 0;
-//   padding: 0 8px;
-//   font-size: 10px;
-//   line-height: 16px;
-//   border-radius: 8px;
-//   color: ${Colors.White};
-//   background-color: ${Colors.Blue[200]};
-//   text-transform: uppercase;
-// `
-
-const AccountName = styled.h5`
-  grid-area: accountname;
-  margin: 0;
-  padding: 0;
-  font-size: 16px;
-  line-height: 24px;
-  font-weight: 700;
-  color: ${Colors.Black[900]};
-`
-
-const AccountCopyAddress = styled.div`
-  display: flex;
-  grid-area: accountaddress;
-  color: ${Colors.Black[400]};
-`
-
-const AccountAddress = styled.span`
-  max-width: 152px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  font-size: 12px;
-  line-height: 18px;
-  color: ${Colors.Black[400]};
-`
-
-const AccountCopyButton = styled(CopyButton)`
-  color: ${Colors.Black[400]};
 `
 
 const AccountBalance = styled.p`
